Make router modular [WTEL-6423]
