+++ src/modules/start-page/router/internals/start-page-route-names.enum.ts
@@ -1,0 +1,5 @@
+enum RouteNames {
+  TheStartPage = 'the-start-page',
+}
+
+export default RouteNames;
