+++ src/modules/start-page/router/internals/start-page-route-paths.enum.ts
@@ -1,0 +1,5 @@
+enum RoutePaths {
+  TheStartPage = '/start-page',
+}
+
+export default RoutePaths;
