+++ src/vite-env.d.ts
@@ -1,0 +1,2 @@
+// https://vite.dev/guide/env-and-mode#intellisense-for-typescript
+/// <reference types="vite/client" />
