+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "esnext",
     "module": "esnext",
-    "lib": ["esnext"],
+    "lib": ["esnext", "DOM"],
     "plugins": [{ "name": "typescript-plugin-css-modules" }],
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
